requirements.txt
flask
langchain-community
sentence-transformers
numpy
faiss-cpu